% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_densmap.R
\name{add_densmap}
\alias{add_densmap}
\title{run densMAP and store the result in reducedDim(x, "DENSMAP")}
\usage{
add_densmap(x, nc = 3L, rdn = NULL, ...)
}
\arguments{
\item{x}{a SingleCellExperiment or a slice of one}

\item{nc}{how many components to keep for the densmap embedding (6)}

\item{rdn}{reducedDimName to use ("HARMONY" if present, "PCA" otherwise)}

\item{...}{other parameters to pass to \code{densvis::densmap}}
}
\value{
\preformatted{   a SingleCellExperiment with reducedDim(x, "DENSMAP")
}
}
\description{
run densMAP and store the result in reducedDim(x, "DENSMAP")
}
\seealso{
\preformatted{  densmap
}
}
