% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plate_metrics.R
\name{add_plate_metrics}
\alias{add_plate_metrics}
\title{some metrics than have come in handy for plate-seq experiments}
\usage{
add_plate_metrics(txis, ...)
}
\arguments{
\item{txis}{a SingleCellExperiment, usually from import_plate_txis}

\item{...}{other parameters, currently just passed to izar_transform}
}
\value{
\preformatted{     a SingleCellExperiment with additional metrics in metadata
}
}
\description{
Prerequisite: annotate the damned gene symbols (with e.g. get_gencode_genes)
}
\seealso{
\preformatted{    make_plate_plot
}
}
