% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_static_trace.R
\name{add_static_trace}
\alias{add_static_trace}
\title{Refactoring plot_velo into multiple functions moving forwards}
\usage{
add_static_trace(txis, embed = "UMAP", colr = NULL, p = NULL, ...)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{embed}{which reducedDims() element to use for plotting? (UMAP)}

\item{colr}{column(s) to use for colors (default: colLabels(txis))}

\item{p}{an existing plotly object to add the trace to (NULL)}

\item{...}{optional arguments to pass to plotly::plot_ly()}
}
\value{
\preformatted{       a plotly trace
}
}
\description{
Refactoring plot_velo into multiple functions moving forwards
}
