% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tpm.R
\name{add_tpm}
\alias{add_tpm}
\title{partition transcripts per million between spliced and unspliced in drop-seq}
\usage{
add_tpm(txis)
}
\arguments{
\item{txis}{SingleCellExperiment with \code{spliced} and \code{unspliced} assays}
}
\value{
\preformatted{     a SingleCellExperiment with spliced_tpm and unspliced_tpm
}
}
\description{
Warning: this is notorious for crashing R sessions when run on droplet data.
Warning number two: don't run this on plate-seq data (UMIed or otherwise)!
FIXME: use the quick skip-all-0-rows approach to make this go reliably.
}
