% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_umap_embedding.R
\name{add_umap_embedding}
\alias{add_umap_embedding}
\title{run UMAP with \code{uwot} and store the embedding model in metadata(x)$uwot}
\usage{
add_umap_embedding(x, nc = 3L, ret = c("sce", "uwot"), rdn = NULL, ...)
}
\arguments{
\item{x}{a SingleCellExperiment or a slice of one}

\item{nc}{how many components to keep for \code{uwot}'s UMAP embedding (3)}

\item{ret}{return the modified "sce" (default) or "uwot" model? ("sce")}

\item{rdn}{reducedDimName to use ("HARMONY" if present, "PCA" otherwise)}

\item{...}{additional parameters for uwot::umap (e.g. \code{y} for SDR)}
}
\value{
\preformatted{ depending on `ret`, an embellished sce or a uwot model
}
}
\description{
run UMAP with \code{uwot} and store the embedding model in metadata(x)$uwot
}
\seealso{
\preformatted{uwot::umap
}
}
