% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_velo_trace.R
\name{add_velo_trace}
\alias{add_velo_trace}
\title{Refactoring plot_velo into multiple functions moving forwards}
\usage{
add_velo_trace(
  txis,
  embed = "UMAP",
  replace = FALSE,
  colr = "velocity_pseudotime",
  p = NULL,
  ...
)
}
\arguments{
\item{txis}{a SingleCellExperiment with metadata(txis)$scVelo}

\item{embed}{which reducedDims() element to use for plotting? (UMAP)}

\item{replace}{replace any existing metadata element \code{embedded}? (FALSE)}

\item{colr}{column(s) to use for colors (default: velocity_pseudotime)}

\item{p}{an existing plotly object to add the trace to (NULL)}

\item{...}{optional arguments to pass to plotly::plot_ly()}
}
\value{
\preformatted{       a plotly trace
}
}
\description{
Refactoring plot_velo into multiple functions moving forwards
}
