% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_seqs_from_hdf5.R
\name{adt_seqs_from_hdf5}
\alias{adt_seqs_from_hdf5}
\title{dump ADT/HTO sequences from an h5 mRNA+ADT file for \link{re}indexing}
\usage{
adt_seqs_from_hdf5(h5_fname, fa_fname = "ADTs.fa", base = "matrix")
}
\arguments{
\item{h5_fname}{the h5 filename (required)}

\item{fa_fname}{the fa filename to write ("ADTs.fa"; NULL elides FASTA)}

\item{base}{the base name group in the h5 file ("matrix")}
}
\value{
\preformatted{       invisibly, a DataFrame of the ADT features and sequences
}
}
\description{
The resulting FASTA can then be fed to Salmon, Kallisto, STAR, etc.
}
\details{
Depending on the dataset, the base name for the features may differ from
the usual "matrix".  This can be detected using rhdf5::h5ls(h5_fname).
}
