% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adts_from_hdf5.R
\name{adts_from_hdf5}
\alias{adts_from_hdf5}
\title{grab ADT/HTO matrices from an h5 mRNA+ADT file}
\usage{
adts_from_hdf5(h5_fname, base = "matrix", stub = NULL)
}
\arguments{
\item{h5_fname}{the h5 filename (required)}

\item{base}{the base name group in the h5 file ("matrix")}

\item{stub}{a column name stub, if using one (NULL)}
}
\value{
\preformatted{       a matrix (i.e., as.matrix(subset of a TENxMatrix))
}
}
\description{
grab ADT/HTO matrices from an h5 mRNA+ADT file
}
