% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotypes.R
\name{biotypes}
\alias{biotypes}
\alias{gene_biotype}
\alias{transcript_biotype}
\alias{gene_biotype,SummarizedExperiment-method}
\alias{gene_biotype<-}
\alias{gene_biotype<-,SummarizedExperiment-method}
\alias{transcript_biotype,SummarizedExperiment-method}
\alias{transcript_biotype<-}
\alias{transcript_biotype<-,SummarizedExperiment-method}
\title{Get or set rowData(object)\link{, biotype} columns in a SummarizedExperiment}
\usage{
gene_biotype(object, ...)

\S4method{gene_biotype}{SummarizedExperiment}(object, ...)

gene_biotype(object, ...) <- value

\S4method{gene_biotype}{SummarizedExperiment}(object, ...) <- value

transcript_biotype(object, ...)

\S4method{transcript_biotype}{SummarizedExperiment}(object, ...)

transcript_biotype(object, ...) <- value

\S4method{transcript_biotype}{SummarizedExperiment}(object, ...) <- value
}
\arguments{
\item{object}{a SummarizedExperiment}

\item{...}{other arguments, as appropriate (which they rarely are)}

\item{value}{value for rowData(object)\link{, biotype} (replaces existing)}
}
\value{
\preformatted{     result of getting/setting rowData(object)[, biotype]
}
}
\description{
Get or set rowData(object)\link{, biotype} columns in a SummarizedExperiment
}
\details{
GET_RD and SET_RD are aped from SingleCellExperiment.
}
