% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnmf.R
\name{bnmf}
\alias{bnmf}
\alias{bnmf,ANY-method}
\alias{bnmf,SummarizedExperiment-method}
\alias{bnmf,SingleCellExperiment-method}
\title{Bayesian NMF and ARD with \pkg{SignatureAnalyzer}}
\usage{
bnmf(x, ...)

\S4method{bnmf}{ANY}(x)

\S4method{bnmf}{SummarizedExperiment}(x, ...)

\S4method{bnmf}{SingleCellExperiment}(x, ...)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A list containing two matrices (W and H) is returned.
}
\description{
Use the fast, scalable, GPU-aware Bayesian NMF implementation in the Broad's
SignatureAnalyzer to factorize matrices and tensors for various purposes.
}
\details{
This function uses the \pkg{SignatureAnalyzer} Python package (found at
\url{https://pypi.org/project/signatureanalyzer/}) for ARD and Bayesian NMF.
This is done via the \pkg{basilisk} package.
}
\author{
Tim Triche, Jr.
}
