% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellrank.R
\name{cellrank}
\alias{cellrank}
\alias{cellrank,SingleCellExperiment-method}
\title{Cell trajectories with \pkg{CellRank}}
\usage{
cellrank(x, ...)

\S4method{cellrank}{SingleCellExperiment}(x, ...)
}
\arguments{
\item{x}{The output of velociraptor::scvelo().}
}
\value{
A \linkS4class{SingleCellExperiment} is returned containing the output.
}
\description{
Use RNA velocity to find root or final states via the \pkg{CellRank} package.
Most of this is a straight-up ripoff of velociraptor so that I can learn how
to operate Python via basilisk for CellRank and SignatureFinder-GPU to run.
}
\details{
This function uses the \pkg{CellRank} Python package (which can be found at
\url{https://pypi.org/project/cellrank/}) for RNA velocity-pseudotime
informed root and leaf finding when constructing trajectories and the like.
This is done via the \pkg{basilisk} package - see the documentation for that package for trouble-shooting.
}
\examples{
# Using mock data to demonstrate the process:
library(scuttle)
sce1 <- mockSCE()
sce2 <- mockSCE()

spliced <- counts(sce1)
unspliced <- counts(sce2)

mats <- list(X=spliced, spliced=spliced, unspliced=unspliced)
velo <- velociraptor::scvelo(mats)
cellrank <- cellrank(velo)

}
\author{
Tim Triche, Jr.
}
