% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sce.R
\name{check_sce}
\alias{check_sce}
\title{check that an SCE will be suitable for muscat, scvelo, plotting, etc.}
\usage{
check_sce(sce, syms = TRUE, velo = TRUE, muscat = TRUE, ...)
}
\arguments{
\item{sce}{the SingleCellExperiment}

\item{syms}{check that genes/transcripts have unique symbols? (TRUE)}

\item{velo}{check for plot_velo compatibility? (TRUE)}

\item{muscat}{check for muscat compatibility (TRUE)}

\item{...}{other parameters, currently ignored}
}
\value{
\preformatted{   logical: whether all the checks passed
}
}
\description{
check that an SCE will be suitable for muscat, scvelo, plotting, etc.
}
