% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_velo_txis.R
\name{cluster_velo_txis}
\alias{cluster_velo_txis}
\title{stupid-simple Louvain clustering. Doesn't currently care about unspliced.}
\usage{
cluster_velo_txis(txis, k = 20, type = "rank", use.dimred = NULL, ...)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{k}{number of nearest neighbors (20)}

\item{type}{what type of SNNGraph to construct ("rank")}

\item{use.dimred}{which reducedDimension embedding to use (autodetermine)}

\item{...}{optional arguments to pass to igraph::cluster_louvain()}
}
\value{
\preformatted{       Louvain cluster assignments for each cell. 
}
}
\description{
If a reduced dimension representation named "HARMONY" exists, the graph will
be constructed on that, otherwise "PCA" will be used (or added if not found)
unless use.dimred is explicitly specified by the user.
}
\details{
It is not yet clear whether an approach like that taken in \code{Rphenoannoy}
would allow for faster and therefore bootstrap-able clustering, and this is
something that will be of interest going forwards. In some cases, using
\code{rank} or \code{jaccard} distances with Louvain clustering produces overly fine-
grained clusters, to the extent that something like SPADE may be needed to
reconcile the results. Forthcoming releases will explore this.
}
