% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_AB.R
\name{diff_AB}
\alias{diff_AB}
\title{this is kind of a helper function, but useful enough to split out}
\usage{
diff_AB(A, B, conf = 0.75, steps = 1, what = "counts", stubs = c("A", "B"))
}
\arguments{
\item{A}{a GRanges}

\item{B}{a GRanges}

\item{conf}{what quantile to use for confidence intervals (.75, quartiles)}

\item{steps}{how many bins to use for the weighted smoother (1; don't smooth)}

\item{what}{what assay fed the original A and B calls? (counts; see Details)}

\item{stubs}{what stubs to use for the difference variables? (c("A","B"))}
}
\value{
\preformatted{ a GRanges with differences between A and B and damped bounds
}
}
\description{
Given compartment estimates A and B, compute estimates of their differences.
}
\details{
\code{what} is left open for future use comparing unspliced/spliced ratios. It
has no effect upon the execution of this function at the moment. Note that
if using the weighted smoother (\code{steps} > 1), then \code{steps} MUST be odd.
}
