% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_doublets.R
\name{drop_doublets}
\alias{drop_doublets}
\title{simple automated doublet marking for pipelining}
\usage{
drop_doublets(txis, ...)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{...}{other arguments, currently unused}
}
\value{
\preformatted{ if successful, a SingleCellExperiment with doublets dropped
}
}
\description{
simple automated doublet marking for pipelining
}
