% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_uninformative_cells.R
\name{drop_uninformative_cells}
\alias{drop_uninformative_cells}
\title{drop cells with an unspliced size factor of 0 (will crash scvelo).}
\usage{
drop_uninformative_cells(txis)
}
\arguments{
\item{txis}{a SingleCellExperiment}
}
\value{
\preformatted{         a SingleCellExperiment with any useless cells omitted
}
}
\description{
drop cells with an unspliced size factor of 0 (will crash scvelo).
}
