% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ensembl.R
\name{fetch_ensembl}
\alias{fetch_ensembl}
\alias{get_ensembl_transcripts}
\alias{get_ensembl_genes}
\title{convenience functions for fetching ENSEMBL/Gencode annotations (genes & txs)}
\usage{
fetch_ensembl(version, species)
}
\arguments{
\item{version}{the ENSEMBL (preferred) or Gencode (as gencodeXYZ) version}

\item{species}{the species to fetch (default is "Homo sapiens")}
}
\value{
\preformatted{     ensdb object (fetch_ensembl) or GRanges[List] of genes/[txs]
}
}
\description{
Gencode genes are based on ENSEMBL builds. When AnnotationHub's version of
Gencode lags behind current (as often happens), it is possible to retrieve
ENSEMBL builds instead. For example, Gencode 33 is ENSEMBL 99; GENCODE 35 is
ENSEMBL 101; and so forth. Moreover, given an ENSEMBL build and a species,
a minimum subset of information (gene id, gene name, gene biotype) is easily
retrieved for annotation purposes. So by default we just grab ENSEMBL genes.
}
\examples{
  fetch_ensembl("101", "Homo sapiens")
  get_ensembl_transcripts("101", "Mus musculus")
  get_ensembl_genes("gencode33")

}
