% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_doublets.R, R/find_doublets.R
\name{drop_doublets}
\alias{drop_doublets}
\alias{find_doublets}
\title{simple automated doublet marking for pipelining, refactored a little bit}
\usage{
drop_doublets(txis, score = NULL, clust = TRUE, ...)

find_doublets(txis, score = NULL, clust = TRUE, ...)
}
\arguments{
\item{txis}{a SingleCellExperiment, ideally with sample and cluster labels}

\item{score}{scoring method ("xgb" if possible, "weighted" otherwise)}

\item{clust}{use clusters? (default is TRUE; set to false if many samples)}

\item{...}{more params for scDblFinder, e.g. \code{trajectoryMode} or \code{BPPARAM}}
}
\value{
\preformatted{ a SingleCellExperiment with doublets marked (or dropped)
}
}
\description{
Note: if colLabels(txis) exists, it will be used to supply cluster labels.
Note: if xgboost is installed, we use "xgb", else "weighted" for the score.
}
\details{
\preformatted{drop_doublets drops doublets. find_doublets just marks them. 
}
}
