% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eligible_cells.R
\name{find_eligible_cells}
\alias{find_eligible_cells}
\title{Find cells that meet certain cluster-wise criteria}
\usage{
find_eligible_cells(txis, mincells = 10)
}
\arguments{
\item{txis}{SingleCellExperiment where !is.null(colLabels(txis))}

\item{mincells}{min cells per cluster per sample (see Details) (10)}
}
\description{
By default, to be eligible, a cell must come from a sample whose presence
in a cluster appears to be non-noise (i.e. not the "soup" part of a mixture),
and after pruning out samples that are all-one-cluster (likely artifacts)
as well as clusters that are all-one-sample (likely artifacts), the cluster
from which the cells come must have at least \code{mincells} cells left in it.
The default for \code{mincells} is 10, so this isn't as harsh as it might seem,
but it's not something you want to be relying upon for SmartSeq-like runs.
One useful application of find_eligible_cells is to feed plot_eligible_cells,
which can help determine whether to raise or lower \code{maxcells} and \code{mincells}
in downsample_txis and label_cells.
}
