% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_cells.R
\name{fix_cells}
\alias{fix_cells}
\title{fix or drop cells with an size factor of 0 (which will crash scvelo).}
\usage{
fix_cells(txis, clean = TRUE)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{clean}{fix negative size factors with cleanSizeFactors? (FALSE)}
}
\value{
\preformatted{         a SingleCellExperiment with any useless cells omitted
}
}
\description{
Also attempts to use scuttle::cleanSizeFactors to recover dubious cells,
if NumGenesExpressed is a column in colData(txis).  This may or may not be
the world's greatest idea
}
