% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_gencode.R
\name{fix_gencode}
\alias{fix_gencode}
\title{helper function to make Gencode version decoding slightly less hellish.}
\usage{
fix_gencode(release)
}
\arguments{
\item{release}{the release string}
}
\value{
\preformatted{     a standardized release string ("33" or "M24" or the like)
}
}
\description{
helper function to make Gencode version decoding slightly less hellish.
}
