% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_rownames.R
\name{fix_rownames}
\alias{fix_rownames}
\title{strip silly ENSEMBL gene/transcript version IDs}
\usage{
fix_rownames(txis, stub = "^ENS", sep = "\\\\.", idx = 1)
}
\arguments{
\item{txis}{an SCE}

\item{stub}{optional leader stub ("^ENS")}

\item{sep}{optional separator ("\\.")}

\item{idx}{which fragment to return (1)}
}
\value{
\preformatted{   same SCE but without goofy rownames 
}
}
\description{
tired of typing this out
}
