% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AB.R
\name{get_AB}
\alias{get_AB}
\title{a shim function}
\usage{
get_AB(x, asy = "counts", chr = "chr3", res = 1e+05, boot = 10, ...)
}
\arguments{
\item{x}{something that inherits from RangedSummarizedExperiment}

\item{asy}{the name of the assay to summarize down ("counts")}

\item{chr}{what chromosome to work on ("chr3")}

\item{res}{resolution for compartmap bins, in base pairs (1e5)}

\item{boot}{how many bootstraps to run for compartmap (10)}

\item{...}{parameters to feed to getATACABsignal (e.g. group, cores, etc)}
}
\value{
\preformatted{   AB signal
}
}
\description{
binarize & TF-IDF SCE, then call boundaries
requires a genome for the SCE or it will fail
requires coordinates for the SCE or it will fail
don't run this on thousands of cells ungrouped or you'll be sorry
}
\seealso{
\preformatted{  compartmap::getATACABsignal
}
}
