% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dupe_txis.R
\name{get_dupe_txis}
\alias{get_dupe_txis}
\title{Mark duplicated mcols(txis)$symbol in a SingleCellExperiment object \code{txis}}
\usage{
get_dupe_txis(txis, mincount = 1, mincells = 1)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{mincount}{minimum count to register as nonzero (1)}

\item{mincells}{how many cells must have at least mincount to retain? (1)}
}
\value{
\preformatted{       data.frame with symbol, ensembl, index, cutoff, cells
}
}
\description{
For large experiments, it may not be possible to run rowsum on a Matrix in
limited memory. And, on occasion, it may not be necessary -- if only one of
the ENSEMBL gene IDs mapping to an HGNC or mouse symbol is nonzero, we don't
need to sum them at all; dropping the near-zero rows is sufficient. This
function allows looping through the ENSEMBL IDs for each "duplicated" symbol
and determining whether rowsum() needs to be called at all.
}
\details{
If there are no duplicated symbols, an empty data.frame results.
}
