% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensembl_genes.R
\name{get_ensembl_genes}
\alias{get_ensembl_genes}
\title{Simple wrapper to annotate genes/transcripts with symbols and so forth}
\usage{
get_ensembl_genes(version, species = "Homo sapiens")
}
\arguments{
\item{version}{ENSEMBL version (from tximeta or gencode mapping)}

\item{species}{ENSEMBL species (from tximeta or gencode release)}
}
\value{
\preformatted{     a GRanges of genes with symbols and biotypes
}
}
\description{
Simple wrapper to annotate genes/transcripts with symbols and so forth
}
