% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensembl_transcripts.R
\name{get_ensembl_transcripts}
\alias{get_ensembl_transcripts}
\title{Get transcripts grouped by gene name.}
\usage{
get_ensembl_transcripts(version, species = "Homo sapiens")
}
\arguments{
\item{version}{the ENSEMBL release version}

\item{species}{the ENSEMBL species ("Homo sapiens")}
}
\value{
\preformatted{       a GRangesList of transcripts grouped by gene
}
}
\description{
Get transcripts grouped by gene name.
}
