% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gencode_ensembl.R
\name{get_gencode_ensembl}
\alias{get_gencode_ensembl}
\title{Gencode releases do not map 1:1 onto ENSEMBL releases, hence this function.}
\usage{
get_gencode_ensembl(release)
}
\arguments{
\item{release}{the Gencode release (e.g. M24 or 30)}
}
\value{
\preformatted{     the ENSEMBL version (e.g. 99 or 96)
}
}
\description{
Gencode releases do not map 1:1 onto ENSEMBL releases, hence this function.
}
