% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gencode_genes.R
\name{get_gencode_genes}
\alias{get_gencode_genes}
\title{A wrapper for get_ensembl_genes, to handle Gencode release mapping}
\usage{
get_gencode_genes(release)
}
\arguments{
\item{release}{GENCODE release (can be got from tximeta json)}
}
\value{
\preformatted{      a GRanges of genes with symbols and biotypes
}
}
\description{
Warning: this function is horrifically poorly tested
}
