% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gencode_species.R
\name{get_gencode_species}
\alias{get_gencode_species}
\title{Gencode prefaces mouse releases with an M, and human releases with nothing.}
\usage{
get_gencode_species(release)
}
\arguments{
\item{release}{the release name (e.g. M24 or 33)}
}
\value{
\preformatted{     the release species (Mus musculus or Homo sapiens)
}
}
\description{
Gencode prefaces mouse releases with an M, and human releases with nothing.
}
