% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gencode_transcripts.R
\name{get_gencode_transcripts}
\alias{get_gencode_transcripts}
\title{Just like get_ensembl_transcripts, but for gencode.}
\usage{
get_gencode_transcripts(release)
}
\arguments{
\item{release}{the GENCODE release string}
}
\value{
\preformatted{       a GRangesList of transcripts grouped by gene
}
}
\description{
Just like get_ensembl_transcripts, but for gencode.
}
