% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_from_barcode.R
\name{get_sample_from_barcode}
\alias{get_sample_from_barcode}
\title{extract the sample/batch name for each cell from a SAMPLENAME_BARCODE scheme}
\usage{
get_sample_from_barcode(txis, sep = "_")
}
\arguments{
\item{txis}{a SingleCellExperiment with colnames(txis) like FOO_BAR_AATGCT}

\item{sep}{the separator character used to paste the column names ("_")}
}
\value{
\preformatted{ a character vector of sample names, e.g. FOO_BAR
}
}
\description{
This function simply assumes that process_velo_txis has renamed the cell
barcodes in a SingleCellExperiment to indicate where they came from, and
pops the last element of the split vector off the end.  This is handy for
automatically running Harmony as a first pass to get rid of batch effects,
though the right way to handle that task while respecting spliced-unspliced
relationships is currently an open research problem.
}
