% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_droplet_txis.R
\name{import_droplet_txis}
\alias{import_droplet_txis}
\title{import droplet RNAseq data (usually from Alevin) into a SingleCellExperiment}
\usage{
import_droplet_txis(
  quants,
  feats = NULL,
  type = c("alevin"),
  QC = TRUE,
  tpms = FALSE,
  su_tpms = FALSE,
  sep = "_",
  fixrn = TRUE,
  ...
)
}
\arguments{
\item{quants}{Directories containing droplet quantification results}

\item{feats}{optional file with intron-exon-gene mappings (guess)}

\item{type}{What type of quantifications are these? ("alevin")}

\item{QC}{Perform rudimentary quality control? (TRUE)}

\item{tpms}{compute TPMs? (FALSE; can create a CHOLMOD error)}

\item{su_tpms}{compute spliced/unspliced TPMs? (FALSE; as above)}

\item{sep}{What string separates sample name from barcode? ("_")}

\item{fixrn}{Fix goofy versioned ENSEMBL gene names? (TRUE)}

\item{...}{additional parameters to pass to tximport, if any}
}
\value{
\preformatted{     A SingleCellExperiment with 'spliced' & 'unspliced' assays.
}
}
\description{
This function is usually called by process_velo_txis(). \code{quants} is, like in
import_plate_txis, a bunch of directories with quantifications in them. \code{QC}
does not exist in import_plate_txis, but it is a flag to perform rudimentary
quality control for droplet data (and is enabled by default). \code{sep} is the
separator character between the assigned sample name and the trimmed barcode.
}
\details{
FIXME: Add Kallisto support and Arkas style txome/repeatome/spikeome support.

TPM calculation is disabled by default due to a tendency to crash sessions.
The function velocessor::add_tpm handles the grunt work for this at present.
It would probably be a good idea to do it more efficiently (e.g. in C++).
}
