% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_kallisto_txis.R
\name{import_kallisto_txis}
\alias{import_kallisto_txis}
\title{import plate-seq runs from Kallisto into a SingleCellExperiment}
\usage{
import_kallisto_txis(
  runs,
  what = c(tpm = "tpm", counts = "est_counts", eff_length = "eff_length"),
  altExps = TRUE,
  asgenes = FALSE,
  t2g = NULL,
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{runs}{a vector of .tsv file paths (usually abundance.tsv) w/names}

\item{what}{columns to import (default: tpm, est_counts, eff_length)}

\item{altExps}{split spliced, unspliced, spikes, viruses & repeats? (TRUE)}

\item{asgenes}{collapse transcripts to genes? (implies altExps & t2g)}

\item{t2g}{data.frame with tx2gene mappings (required if asgenes)}

\item{BPPARAM}{optional BiocParallel parameter object for bplapply()}

\item{...}{additional parameters to pass to .split_altExps}
}
\value{
\preformatted{     A SingleCellExperiment
}
}
\description{
Automated (more or less) importation of STORM-seq and SMART-seq\link{123} data
for downstream attempts at merging, index visualization, etc.  Assumes that
users will quantify against ENSEMBL transcripts instead of ridiculous crap
like GENCODE and similar. Default is ERCC spikes, ENSembl transcripts, and
if found, intronic quantifications of the same. Everything else gets labeled
as 'viruses_and_repeats' since in our transcriptomes, that's what it is. If
altExps is set to FALSE, no splitting will be attempted.
}
\examples{

t2g <- read.delim("../hs_mm_spikes_repeats_ens100.trimmed.velo.tx2gene.tsv",
                  sep="\t", row.names=1, head=FALSE)
t2g <- t2g[grep("ENS", rownames(t2g)), , drop=FALSE]

}
