% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plate_txis.R
\name{import_plate_txis}
\alias{import_plate_txis}
\title{import plate-seq velocity information (from Salmon) to a SingleCellExperiment}
\usage{
import_plate_txis(quants, t2g = NA, type = "salmon", ...)
}
\arguments{
\item{quants}{where the quant.sf files are}

\item{t2g}{required file with tx2gene information for TPM calcs}

\item{type}{What type of quantifications are these? ("salmon")}

\item{...}{additional parameters to pass to tximport, if any}
}
\value{
\preformatted{     A SingleCellExperiment with 'spliced' & 'unspliced' assays.
}
}
\description{
import plate-seq velocity information (from Salmon) to a SingleCellExperiment
}
