% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_infercnv.R
\name{infercnv}
\alias{infercnv}
\title{massage a (possibly velocitized) SingleCellExperiment into an infercnv object}
\usage{
infercnv(x, anno_col = "status", ref_groups = c("normal"), ...)
}
\arguments{
\item{x}{a SingleCellExperiment}

\item{anno_col}{the colData column containing cell annotations ("status")}

\item{ref_groups}{annotations of the cells to use as references ("normal")}

\item{...}{any other arguments for infercnv::CreateInfercnvObject}
}
\value{
\preformatted{       an infercnv object
}
}
\description{
massage a (possibly velocitized) SingleCellExperiment into an infercnv object
}
\details{
We do not require or import infercnv, because the requirement for JAGS can
be a showstopper for HPC installations. This function will stop() if the
infercnv package namespace cannot be attached, however, for obvious reasons.
I have no idea why the \code{infercnv} function name was not taken by \code{infercnv},
but since it wasn't, I stole it.
}
