% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mtx.R
\name{load_mtx}
\alias{load_mtx}
\title{Load a triple of gene names, cell names, gene x cell counts as from bustools}
\usage{
load_mtx(path, verbose = TRUE)
}
\arguments{
\item{path}{where the counts files are (genes, barcodes, mtx)}

\item{verbose}{be verbose? (TRUE)}
}
\value{
\preformatted{       a dgCMatrix 
}
}
\description{
The returned matrix is a dgCMatrix as used by (e.g.) SingleCellExperiment.
This function is useful when processing data from (e.g.) kallisto|bustools,
for example, CITE-seq data processed with KITE, or when comparing uniquely
mapped vs. multimapped counts, as one might do for repetitive elements.
}
