% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plate_plot.R
\name{make_plate_plot}
\alias{make_plate_plot}
\title{plot outcomes of 96-, 384-, or 1536-well plate sequencing runs}
\usage{
make_plate_plot(txis, column = "NumGenesExpressed", ...)
}
\arguments{
\item{txis}{a SingleCellExperiment with either $well or $row and $column}

\item{column}{which column to use for the data to plot? ($NumGenesExpressed)}

\item{...}{additional arguments for platetools::raw_map}
}
\value{
\preformatted{   a ggplot object 
}
}
\description{
plot outcomes of 96-, 384-, or 1536-well plate sequencing runs
}
\examples{

 # helper fn
 revglue <- function(x) gsub(" ", "", paste(rev(x), collapse=""))
 wells <- toupper(apply(expand.grid(1:24, letters[1:16]), 1, revglue))

 # mock up an SCE
 maxgenes <- 100
 library(scDblFinder) 
 sce <- mockDoubletSCE(ncells=c(184, 150, 50), ngenes=maxgenes)[, 1:384]
 sce$NumGenesExpressed <- colSums(counts(sce) > 0)
 sce$well <- wells

 # plot it with $well
 make_plate_plot(sce)

 # plot it using colnames
 colnames(sce) <- sce$well
 sce$well <- NULL # drop
 # ensure case is ignored for now
 changecase <- sample(seq_len(ncol(sce)), round(ncol(sce)/2))
 colnames(sce)[changecase] <- toupper(colnames(sce)[changecase])
 colnames(sce)[-changecase] <- tolower(colnames(sce)[-changecase])
 make_plate_plot(sce)

}
\seealso{
platetools::raw_map

platetools::b_map
}
