% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spliced_index.R
\name{make_spliced_index}
\alias{make_spliced_index}
\title{aped from https://combine-lab.github.io/alevin-tutorial/2020/alevin-velocity}
\usage{
make_spliced_index(gtf, fa = NULL, flankLength = 90L, verbose = TRUE)
}
\arguments{
\item{gtf}{a GTF file, usually from ENSEMBL}

\item{fa}{a FASTA file, usually a genome (same as .gtf but .fa)}

\item{flankLength}{how many bases of intron (90)}

\item{verbose}{squawk? (TRUE)}
}
\value{
\preformatted{ a list with information about the index and tximeta created
}
}
\description{
Given a fasta and a gtf, make a spliced index out of it.
}
