% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_storm_trimmer.R
\name{make_storm_trimmer}
\alias{make_storm_trimmer}
\title{simple function to emit a trimming script for STORM-seq preps}
\usage{
make_storm_trimmer(stub)
}
\arguments{
\item{stub}{the shared part of the fastq filenames}
}
\description{
simple function to emit a trimming script for STORM-seq preps
}
