% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tximeta.R
\name{make_tximeta}
\alias{make_tximeta}
\title{downstream processor for make_spliced_index results (post-salmon-index)}
\usage{
make_tximeta(params, source = "ENSEMBL", release = "", organism = NULL)
}
\arguments{
\item{params}{the results from make_spliced_index}

\item{source}{where the annotations came from originally (ENSEMBL)}

\item{release}{what release of the transcriptome it is (NULL, don't know)}

\item{organism}{which organism this index is for (guess from genome)}
}
\value{
\preformatted{     a new entry in the bfc and a new linkedTxome
}
}
\description{
downstream processor for make_spliced_index results (post-salmon-index)
}
