% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sces.R
\name{merge_sces}
\alias{merge_sces}
\title{convenience function for merging datasets after QC (e.g. doublet finding)}
\usage{
merge_sces(x, y)
}
\arguments{
\item{x}{the first SCE}

\item{y}{the second SCE}
}
\value{
a merged SCE
}
\description{
This function assumes that harmonized UMAP and maybe scVelo objects exist
(the latter in metadata(x) and/or metadata(y)), but also assumes that the
user may not want to brute-force merge them, so they are saved into the
metadata slot of the new merged object for future reference. scDblFinder
columns in rowRanges are not, however; they cause problems with cbind().
}
