% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonzero.R
\name{nonzero}
\alias{nonzero}
\alias{nz}
\alias{nz_rows}
\alias{nz_rownames}
\alias{nz_byrow}
\alias{nz_cols}
\alias{nz_colnames}
\alias{nz_bycol}
\title{find nonzero entries in a sparse (dgcMatrix) matrix}
\usage{
nonzero(x, agg = NA)

nz(x)

nz_rows(x)

nz_rownames(x)

nz_byrow(x)

nz_cols(x)

nz_colnames(x)

nz_bycol(x)
}
\arguments{
\item{x}{the matrix}
}
\value{
\preformatted{ for `nonzero`, a list with elements "nz", "nrow", and "ncol"
}
}
\description{
find nonzero entries in a sparse (dgcMatrix) matrix
}
\details{
\code{nz} is borrowed from the DAPAR package and is the base function.
\code{nz_rownames} returns the rownames for nonzero rows of \code{x}.
\code{nz_rows} just wraps unique(x@i) for a dgCMatrix \code{x}.
\code{nz_colnames} returns the colnames for nonzero columns of \code{x}.
\code{nz_cols} finds nonzero columns of a dgCMatrix \code{x}.
\code{nonzero} retains additional information about \code{x} alongside \code{nz}.
These are handy for (e.g.) computing TPMs and the like in 3'/5' data.
}
