% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ABC.R
\name{plot_ABC}
\alias{plot_ABC}
\title{plot compartments from samples A & B, perhaps by diffing matrices}
\usage{
plot_ABC(
  A,
  B,
  diff = TRUE,
  with = TRUE,
  what = "counts",
  steps = 1L,
  nameA = "A",
  nameB = "B",
  ...
)
}
\arguments{
\item{A}{a GRanges, usually from compartmap or get_AB}

\item{B}{a GRanges, usually from compartmap or get_AB (but see Details)}

\item{diff}{plot their differences, too? (TRUE, and again, see Details)}

\item{with}{plot with confidence intervals? (TRUE; yet again, see Details)}

\item{what}{what assay fed the original A and B calls? (also, see Details)}

\item{steps}{damping step if using mean smoother -- default is 1, no smooth}

\item{nameA}{name for A in the plots ("A")}

\item{nameB}{name for B in the plots ("B")}

\item{...}{more parameters to feed to diff_AB}

\item{mat}{difference correlation matrices instead? (FALSE, experimental)}
}
\description{
FIXME: use an exponential smoother instead of the mean smoother
}
\details{
There are two nonstandard options here, neither of which should be used just
yet (if it breaks, you get to keep the pieces, but I'm happy to accept PRs).
The first is if \code{what} is set to "ratio" -- the assumption is that we are
looking at "chromatin potential" and, accordingly, the function will assume
that \code{A}, \code{B}, or both were generated from both spliced and unspliced counts.
The second is if both \code{diff} and \code{with} are specified; then the CIs will be
treated as upper and lower bounds for the differences in a third panel. The
behavior of this function is undefined if \code{diff} and \code{with} are TRUE along
with \code{what} being set to "ratio".  The behavior of this function is also
undefined if \code{mat} is set to true, not because it's particularly hard, but
because the difference detection is a work in progress.
}
\seealso{
\preformatted{  compartmap::plotAB
}
}
