% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eligible_cells.R
\name{plot_eligible_cells}
\alias{plot_eligible_cells}
\title{Plot a heatmap of cells x samples that fit certain sampling criteria.}
\usage{
plot_eligible_cells(txis, mincells = 10, logcells = FALSE, ...)
}
\arguments{
\item{txis}{SingleCellExperiment where !is.null(colLabels(txis))}

\item{mincells}{min cells per cluster per sample (see Details) (10)}

\item{logcells}{use log1p(cells) instead of raw cell numbers? (FALSE)}

\item{...}{additional arguments to pass to ComplexHeatmap::Heatmap}
}
\description{
See the documentation for find_eligible_cells regarding these criteria.
This can be a useful diagnostic when setting \code{maxcells} for label_cells.
}
\seealso{
label_cells
}
