% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{plot_features}
\alias{plot_features}
\title{some equivalents to Seurat standards, with or without velocity vectors}
\usage{
plot_features(txis, features, joint = TRUE, pal = "inferno", ...)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{features}{names of features (rows) to plot}

\item{joint}{for multiple features, plot joint density? (TRUE)}

\item{pal}{what palette to use? ("inferno", because it looks nice)}

\item{...}{other arguments for Nebulosa::plot_density}
}
\description{
Mostly for joint density plots via Nebulosa with velocity overlays.
}
