% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_velo.R
\name{plot_velo}
\alias{plot_velo}
\title{Generate interactive 3D UMAP plot with cones for velocity. Work in progress.}
\usage{
plot_velo(txis, ...)
}
\arguments{
\item{txis}{a SingleCellExperiment with metadata(txis)$scVelo}

\item{...}{optional arguments to pass to plotly::plot_ly()}
}
\value{
\preformatted{       a plotly plot
}
}
\description{
The opacity of the cones is manipulated so that dot colors can be seen.
}
