% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_velo_txis.R
\name{process_velo_txis}
\alias{process_velo_txis}
\title{Load some Alevin runs quantified with velocity information}
\usage{
process_velo_txis(
  runs,
  txstub,
  qm = "alevin/quants_mat.gz",
  QC = TRUE,
  HARMONY = FALSE,
  CLUSTER = TRUE,
  DEDUPE = FALSE,
  SCVELO = FALSE,
  meta = FALSE,
  ...,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{runs}{a named vector of paths (if unnamed, you'll have problems)}

\item{txstub}{the shared portion of the transcriptome annotation files}

\item{qm}{the name and path of the Alevin quants matrix file in 'runs'}

\item{QC}{do rudimentary quality control on each Alevin dataset? (TRUE)}

\item{HARMONY}{run Harmony on merged dataset? (FALSE, as harmony is non-BioC)}

\item{CLUSTER}{apply simple Louvain clustering to the result? (TRUE)}

\item{DEDUPE}{de-dupe genes and cells (doublets)? (FALSE unless SCVELO=TRUE)}

\item{SCVELO}{run scVelo on the (perhaps harmonized) result? (FALSE, slow)}

\item{meta}{use tximeta for import_velo_txis? (FALSE; avoid bfc issues)}

\item{...}{other arguments to pass to velociraptor::scvelo, if any}

\item{BPPARAM}{if running in parallel, provide a MulticoreParam or similar}
}
\value{
\preformatted{   a SingleCellExperiment with 
}
}
\description{
This is evolving rapidly to handle microwell-seq and plate-seq data.
}
\details{
The results of this function can be further QC'ed and/or batch corrected,
for example with Harmony or sva::ComBat, then passed to velociraptor::scvelo
and/or velocessor::cellrank for data-driven cell fate inference. We are
investigating various factorization approaches that may allow for sensible
batch effect correction while respecting the relationships between spliced
and unspliced transcript abundance, but have not yet settled on a default.

See also plot_velo for a relatively straightforward exploratory plot
based on UMAP coordinates and the downsampled subset of cells from the
SingleCellExperiment that this function and velociraptor::scvelo yields.
}
