% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_onto_umap.R
\name{project_onto_umap}
\alias{project_onto_umap}
\title{given a \code{uwot} model in metadata(x)$uwot, project y onto those coordinates}
\usage{
project_onto_umap(x, y, rdn = NULL, ...)
}
\arguments{
\item{x}{a SingleCellExperiment where !is.null(metadata(x)$uwot)}

\item{y}{a SingleCellExperiment with samples to map onto x's model}

\item{rdn}{reducedDimName to use ("HARMONY" if present, "PCA" otherwise)}

\item{...}{additional parameters for uwot::umap_transform}
}
\value{
\preformatted{ embedding coordinates for y onto x
}
}
\description{
given a \code{uwot} model in metadata(x)$uwot, project y onto those coordinates
}
\seealso{
\preformatted{uwot::umap_transform
}
}
