% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_sequentially.R
\name{relevel_sequentially}
\alias{relevel_sequentially}
\title{convenience function for reordering factors without thinking very hard}
\usage{
relevel_sequentially(x, lvls)
}
\arguments{
\item{x}{a factor, or something that can become one}

\item{lvls}{a vector with level names that include all levels of x}
}
\value{
\preformatted{ a releveled factor version of x
}
}
\description{
I wrote this a million years ago for TARGET and have used it ever since.
It's not terribly clever but it can help A LOT with plotting to reorder.
}
