% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_dimred.R
\name{rescale_dimred}
\alias{rescale_dimred}
\title{rescale a reduced-dimensional representation for plotting purposes}
\usage{
rescale_dimred(txis, dimred = "UMAP")
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{dimred}{name of the existing reduced dimensional representation}
}
\value{
\preformatted{       a SingleCellExperiment with a paste0("scaled_", dimred)
}
}
\description{
Project the quantiles of a given representation onto a Gaussian and use that.
}
