% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_clusters.R
\name{sample_clusters}
\alias{sample_clusters}
\title{downsample by clusters for NMF/NTF exploration}
\usage{
sample_clusters(txis, n = 100)
}
\arguments{
\item{txis}{a SingleCellExperiment with $cluster set}

\item{n}{how many cells per cluster (default is 100)}
}
\value{
\preformatted{ a downsampled SingleCellExperiment
}
}
\description{
downsample by clusters for NMF/NTF exploration
}
