% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.R
\name{scdb}
\alias{scdb}
\title{automates fetching of Valentine Svenssen's single-cell database}
\usage{
scdb(URL = "http://www.nxn.se/single-cell-studies/data.tsv", ...)
}
\arguments{
\item{URL}{the URL if it changes}

\item{...}{anything else to tell utils::read.delim}
}
\value{
\preformatted{ a data.frame, with the Date column as an actual Date
}
}
\description{
scdb() tidies up the database a bit, with Date, URL, Cells, and Velocity
}
\details{
The database itself lives at http://www.nxn.se/single-cell-studies/data.tsv
}
\examples{

  db <- scdb()

  library(useful)
  library(tidyverse)
  subset(db, Velocity == TRUE) \%>\% corner

  columns <- c("Title", "Data.location", "URL")
  subset(db, Tissue == "Blood" & Cells > 50000)[, columns]
  subset(db, grepl("CITE", Technique))[, columns]
  subset(db, grepl("Amit", Authors))[, columns]

  a_year_ago <- (Sys.Date() - 365)
  subset(db, Date > a_year_ago) \%>\% corner

  five_years_ago <- (Sys.Date() - (5*365.25))
  p1 <- ggplot(subset(db, Date > five_years_ago), 
               aes(x = Date, fill = Velocity)) + 
    geom_histogram() + 
    scale_x_date() + 
    xlab("year") + 
    scale_y_log10() + 
    ylab("papers") + 
    NULL 

  p1 + theme_minimal() + ggtitle("Velocity in papers")

  p2 <- ggplot(subset(db, 
                      Organism \%in\% c("Human", "Mouse", "Zebrafish") & 
                      Date > five_years_ago), 
               aes(x=Date, y=Cells, color=Organism, group=Organism)) + 
    geom_point() + 
    geom_smooth() + 
    scale_y_log10() + 
    NULL

  p2 + facet_grid(vars(Organism)) + theme_minimal() 

}
\references{
Svensson V, Beltrame EdV, Pachter L (2020), A curated database reveals
trends in single-cell transcriptomics, Database, 28 November 2020
https://doi.org/10.1093/database/baaa073
}
