% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_SE.R
\name{split_SE}
\alias{split_SE}
\title{convenience function: split a \link{Summarized/SingleCell}Experiment by columns}
\usage{
split_SE(sce, splt)
}
\arguments{
\item{sce}{the SE-like object}

\item{splt}{the factor to split by, or a colData column in sce}
}
\value{
\preformatted{ a list of SE or SCE objects, one per level of splt
}
}
\description{
convenience function: split a \link{Summarized/SingleCell}Experiment by columns
}
