% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veload.R
\name{veload}
\alias{veload}
\title{world's dumbest velocity loader, currently only for Alevin output}
\usage{
veload(dir = ".", suffix = "_alevin", ...)
}
\arguments{
\item{dir}{where the output folders are (".")}

\item{suffix}{what the suffix of an output folder is ("_alevin")}
}
\description{
world's dumbest velocity loader, currently only for Alevin output
}
