% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_assays.R
\name{velocity_assays}
\alias{velocity_assays}
\alias{unspliced}
\alias{spliced}
\alias{spliced,SingleCellExperiment-method}
\alias{spliced<-}
\alias{spliced<-,SingleCellExperiment-method}
\alias{unspliced,SingleCellExperiment-method}
\alias{unspliced<-}
\alias{unspliced<-,SingleCellExperiment-method}
\title{Get or set assays particular to velocitized SingleCellExperiments}
\usage{
spliced(object, ...)

\S4method{spliced}{SingleCellExperiment}(object, ...)

spliced(object, ...) <- value

\S4method{spliced}{SingleCellExperiment}(object, ...) <- value

unspliced(object, ...)

\S4method{unspliced}{SingleCellExperiment}(object, ...)

unspliced(object, ...) <- value

\S4method{unspliced}{SingleCellExperiment}(object, ...) <- value
}
\arguments{
\item{object}{a SingleCellExperiment}

\item{...}{other arguments, as appropriate (which they rarely are)}

\item{value}{the value for the assay in question}
}
\value{
\preformatted{     the result of getting or setting the assay
}
}
\description{
Specifically, \code{spliced} and \code{unspliced} are made into generics.
}
\details{
GET and SET are stolen directly from SingleCellExperiment.
}
\seealso{
SingleCellExperiment
}
