% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_metadata.R
\name{velocity_metadata}
\alias{velocity_metadata}
\alias{embedded}
\alias{scVelo}
\alias{scVelo,SingleCellExperiment-method}
\alias{scVelo<-}
\alias{scVelo<-,SingleCellExperiment-method}
\alias{embedded,SingleCellExperiment-method}
\alias{embedded<-}
\alias{embedded<-,SingleCellExperiment-method}
\title{Get or set metadata items for velocitized SingleCellExperiments}
\usage{
scVelo(object, ...)

\S4method{scVelo}{SingleCellExperiment}(object, ...)

scVelo(object, ...) <- value

\S4method{scVelo}{SingleCellExperiment}(object, ...) <- value

embedded(object, ...)

\S4method{embedded}{SingleCellExperiment}(object, ...)

embedded(object, ...) <- value

\S4method{embedded}{SingleCellExperiment}(object, ...) <- value
}
\arguments{
\item{object}{a SingleCellExperiment}

\item{...}{other arguments, as appropriate (which they rarely are)}

\item{value}{value to replace existing metadata(object)[\link{name}]}
}
\value{
\preformatted{     result of getting or setting metadata(object)[[name]]
}
}
\description{
Specifically, \code{scVelo} and \code{embedded} have their own generics.
}
