% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_velo_txis.R
\name{cluster_velo_txis}
\alias{cluster_velo_txis}
\title{stupid-simple Louvain clustering. Doesn't currently care about unspliced.}
\usage{
cluster_velo_txis(txis, type = "jaccard", use.dimred = NULL, ...)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{type}{what type of SNNGraph to construct ("jaccard")}

\item{use.dimred}{which reducedDimension embedding to use (autodetermine)}

\item{...}{optional arguments to pass to igraph::cluster_louvain()}
}
\value{
\preformatted{       Louvain cluster assignments for each cell. 
}
}
\description{
If a reduced dimension representation named "HARMONY" exists, the graph will
be constructed on that, otherwise "PCA" will be used (or added if not found)
unless use.dimred is explicitly specified by the user.
}
