% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupe_velo_txis.R
\name{dedupe_velo_txis}
\alias{dedupe_velo_txis}
\title{deduplicate symbols in a SingleCellExperiment object with multiple assays}
\usage{
dedupe_velo_txis(txis)
}
\arguments{
\item{txis}{a SingleCellExperiment}
}
\value{
\preformatted{       txis, but with duplicate representations of symbols summed
}
}
\description{
rowData(txis) \emph{must} have columns 'symbol' and 'gene_id' for this to work.
Perhaps not coincidentally, this is what ensembldb returns for a given build.
}
