% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_doublets.R
\name{drop_doublets}
\alias{drop_doublets}
\title{simple automated doublet marking for pipelining}
\usage{
drop_doublets(txis, score = "weighted", clust = TRUE, ...)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{score}{scoring method (default is "weighted", use "xgb" if you can)}

\item{clust}{use clusters? (default is TRUE; set to false with many samples)}

\item{...}{more params for scDblFinder, e.g. \code{trajectoryMode} or \code{BPPARAM}}
}
\value{
\preformatted{ if successful, a SingleCellExperiment with doublets dropped
}
}
\description{
Note: if colLabels(txis) exists, it will be used to supply cluster labels.
Note: we skip the xgboost option in favor of "weighted" for HPC usage.
}
