% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ensembl.R
\name{fetch_ensembl}
\alias{fetch_ensembl}
\alias{get_ensembl_transcripts}
\alias{get_ensembl_genes}
\title{convenience function for fetching ENSEMBL annotations (genes & txs)}
\usage{
fetch_ensembl(version, species)
}
\arguments{
\item{version}{the ENSEMBL version}

\item{species}{the ENSEMBL species ("Homo sapiens")}
}
\value{
\preformatted{     an ensdb object
}
}
\description{
convenience function for fetching ENSEMBL annotations (genes & txs)
}
\examples{
  fetch_ensembl("101", "Homo sapiens")
  get_ensembl_transcripts("101", "Mus musculus")
  get_gencode_transcripts("m24")
  get_gencode_genes("gencode33")

}
