% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mixtures.R
\name{fit_mixtures}
\alias{fit_mixtures}
\title{Fit mixtures. Turns out this was useful outside of downsampling, so...}
\usage{
fit_mixtures(x, G = 1:2, log = TRUE, keep = FALSE)
}
\arguments{
\item{x}{the thing to fit a 1D mixture to}

\item{G}{vector of component numbers (default is 1:2)}

\item{log}{take log1p of x? (TRUE)}

\item{keep}{keep the model fits as an attribute? (FALSE)}
}
\value{
\preformatted{ list of fits, length 1 if log == FALSE 
}
}
\description{
Fit a Gaussian mixture model to x and log1p(x) with 1-2 components (default).
Whichever has the lowest BIC wins. Nothing fancy! If nothing wins, then put
all of the observations into the same group and leave them all eligible. For
prescreening of potential cluster-by-sample combinations, see find_eligible.
}
\seealso{
\preformatted{find_eligible
}
}
