% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AB.R
\name{get_AB}
\alias{get_AB}
\title{a shim function}
\usage{
get_AB(x, asy = "counts", chr = "chr3", res = 1e+05, ...)
}
\arguments{
\item{x}{something that inherits from RangedSummarizedExperiment}

\item{asy}{the name of the assay to summarize down ("counts")}

\item{chr}{what chromosome to work on ("chr3")}

\item{res}{resolution for compartmap bins, in base pairs (1e5)}

\item{...}{additional parameters to feed to getATACABsignal}
}
\value{
\preformatted{   AB signal 
}
}
\description{
binarize and TF-IDF, then call boundaries
}
