% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_velo_txis.R
\name{harmonize_velo_txis}
\alias{harmonize_velo_txis}
\title{harness for velocity correction (currently ignorant of spliced/unspliced)}
\usage{
harmonize_velo_txis(txis, colname = "sample", how = "default", ...)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{colname}{name of colData column with the batch variable ("sample")}

\item{how}{how to respect spliced/unspliced relation, currently unused}

\item{...}{other arguments to pass to RunHarmony}
}
\value{
\preformatted{     a SingleCellExperiment with "HARMONY" in reducedDimNames
}
}
\description{
Harmony is not a Bioconductor package, so the user needs to install it from
\url{https://github.com/immunogenomics/harmony} prior to using this function.
The UMAP projection will be updated to reflect harmonized coordinates unless
the user sets UMAP to FALSE when calling the function.
}
