% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plate_txis.R
\name{import_plate_txis}
\alias{import_plate_txis}
\title{import plate-seq velocity information (from Salmon) to a SingleCellExperiment}
\usage{
import_plate_txis(
  quants,
  t2g = NA,
  type = "salmon",
  gtf = NULL,
  qc = TRUE,
  fixrn = TRUE,
  ...
)
}
\arguments{
\item{quants}{where the quant.sf files are}

\item{t2g}{required file with tx2gene information for TPM calcs}

\item{type}{What type of quantifications are these? ("salmon")}

\item{gtf}{Where an expanded GTF lives if not collapsed to gene-level}

\item{qc}{Whether QC should be performed to toss low-quality cells}

\item{fixrn}{Fix goofy versioned ENSEMBL gene names? (TRUE)}

\item{...}{additional parameters to pass to tximport, if any}
}
\value{
\preformatted{     A SingleCellExperiment with 'spliced' & 'unspliced' assays.
}
}
\description{
Automated (more or less) importation of STORM-seq and SMART-seq\link{123} data for
downstream attempts at merging with droplet data, index sort visualization,
and so forth.
}
\details{
FIXME: Add Kallisto support and Arkas style txome/repeatome/spikeome support.
FIXME: Add in ability to support bootstraps/Gibbs samples from salmon
}
