% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_velo_txis.R
\name{import_velo_txis}
\alias{import_velo_txis}
\title{import velocity information (usually from Alevin) into a SingleCellExperiment}
\usage{
import_velo_txis(cdat, quant = c("alevin"), meta = FALSE, QC = TRUE, sep = "_")
}
\arguments{
\item{cdat}{A row of the colData prepared by process_velo_txis.}

\item{quant}{What type of quantifications are these? ("alevin")}

\item{meta}{Use tximeta for imports? (FALSE; can fail unexpectedly)}

\item{QC}{Perform rudimentary quality control? (TRUE)}

\item{sep}{What string separates sample name from barcode? ("_")}
}
\value{
\preformatted{     A SingleCellExperiment with 'spliced' & 'unspliced' assays.
}
}
\description{
This function is usually called by process_velo_txis(). The 'cdat' argument
takes a single-row data.frame containing 'names' and 'files' as its columns.
'quant' is the type of quantifications; at present, only alevin is supported.
'sep' allows deduplicating barcodes by sample using cbind without incident
once the results of import_velo_txis are returned to process_velo_txis.
}
