% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercnv.R
\name{infercnv}
\alias{infercnv}
\title{massage a (possibly velocitized) SingleCellExperiment into an infercnv object}
\usage{
infercnv(
  x,
  group_col = "cnv_group",
  ref_prefix = "normal_",
  obs_prefix = "tumor_",
  by_cluster = FALSE,
  downsample = TRUE,
  maxcells = 100,
  run = TRUE,
  cutoff = NULL,
  ...
)
}
\arguments{
\item{x}{a SingleCellExperiment}

\item{group_col}{the colData column with cell annotations ("cnv_group")}

\item{ref_prefix}{prefix for reference sample clusters ("normal_")}

\item{obs_prefix}{prefix for tumor sample clusters ("tumor_")}

\item{by_cluster}{add clusters for references and observations? (FALSE)}

\item{downsample}{downsample? (TRUE, \code{maxcells} per cluster per sample)}

\item{maxcells}{how many cells per sample per cluster to grab? (100)}

\item{run}{perform infercnv run with "sensible" defaults? (TRUE)}

\item{cutoff}{expression cutoff for infercnv::run (default: guess)}

\item{...}{other arguments for infercnv::run, e.g. \code{num_threads}}
}
\value{
\preformatted{         an infercnv object
}
}
\description{
massage a (possibly velocitized) SingleCellExperiment into an infercnv object
}
\details{
We do not require or import infercnv, because the requirement for JAGS can
be a showstopper for HPC installations. This function will stop() if the
infercnv package namespace cannot be attached, however, for obvious reasons.
I have no idea why the \code{infercnv} function name was not taken by \code{infercnv},
but since it wasn't, I stole it. If \code{NumGenesExpressed} is among the colData
columns for the SingleCellExperiment, and \code{run}==TRUE, then this function
will use the former to guess a sensible value for the latter. The velocessor
package also registers a halfway decent \code{show} method for infercnv objects.
}
