% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_cells.R
\name{label_cells}
\alias{label_cells}
\title{barebones cell labeling function}
\usage{
label_cells(
  txis,
  species = NULL,
  ret = c("sce", "labels"),
  downsample = NULL,
  maxcells = 50,
  ...
)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{species}{what species is this from? (autodetect Hs/Mm)}

\item{ret}{what kind of object to return}

\item{downsample}{downsample? (if ncol(txis) > 20000, TRUE, else FALSE)}

\item{maxcells}{maximum number of cells per cluster per sample (50)}

\item{...}{other arguments passed to SingleR}
}
\value{
depending on the value of 'ret', either an SCE or a set of labels
}
\description{
Uses singleR to label celltypes. More sophisticated users may want to work
directly with SingleR input and output for (e.g.) specific gene sets.
}
\details{
Autodetection of species will fail if the genome for the SingleCellExperiment
does not contain "GRCm", "mm", "GRCh", or "hg".
}
