% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_cells.R
\name{label_cells}
\alias{label_cells}
\title{barebones cell labeling function}
\usage{
label_cells(
  txis,
  species = NULL,
  ret = c("sce", "labels"),
  downsample = NULL,
  maxcells = 10,
  mincells = 10,
  label = "label.main",
  ref = NULL,
  ...
)
}
\arguments{
\item{txis}{a SingleCellExperiment}

\item{species}{what species is this from? (autodetect Hs/Mm)}

\item{ret}{what kind of object to return}

\item{downsample}{downsample? (if ncol(txis) > 20000, TRUE, else FALSE)}

\item{maxcells}{maximum number of cells \emph{per_sample} per cluster (20)}

\item{mincells}{minimum number of cells per cluster (10)}

\item{label}{which label to use ("label.main")}

\item{ref}{a reference SummarizedExperiment (default: HPCD/ImmGen)}

\item{...}{other arguments passed to SingleR}
}
\value{
depending on the value of 'ret', either an SCE or a set of labels
}
\description{
Uses singleR to label celltypes. More sophisticated users may want to work
directly with SingleR input and output for (e.g.) specific gene sets.
}
\details{
Autodetection of species will fail if the genome for the SingleCellExperiment
does not contain "GRCm", "mm", "GRCh", or "hg". Once reasonable reference
datasets are available for GRCz/dr genomes, we will support those too.

Twenty cells per sample per cluster may be an awful lot of cells if you have
a large number of samples (i.e. GEMcode preps). Setting \code{maxcells} as low as
10, or perhaps even lower (see upcoming cluster-by-sample plots), if the
cells comprising a cluster come from across many samples. The quick way to
find this out is with a plot,
}
