% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mtx.R
\name{load_mtx}
\alias{load_mtx}
\title{Load a triple of gene names, cell names, gene x cell counts as from bustools}
\usage{
load_mtx(path = ".", verbose = TRUE, frags = NULL, splt = FALSE, spltcol = 3)
}
\arguments{
\item{path}{where the counts files (features, barcodes, mtx) (".")}

\item{verbose}{be verbose? (TRUE)}

\item{frags}{optional patterns to scan for features, barcodes, matrix}

\item{splt}{split by feature type? (FALSE)}

\item{spltcol}{if splitting by feature type, which column to use? (3)}
}
\value{
\preformatted{       a dgCMatrix 
}
}
\description{
The returned matrix is a dgCMatrix as used by (e.g.) SingleCellExperiment.
This function is useful when processing data from (e.g.) kallisto|bustools,
for example, CITE-seq data processed with KITE, or when comparing uniquely
mapped vs. multimapped counts, as one might do for repetitive elements.
}
\details{
This function works fine on gzipped files (tested on a 180K x 2761K matrix).
}
