% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densities.R
\name{plot_densities}
\alias{plot_densities}
\title{another helper that grew into a real live function (albeit a pathetic one)}
\usage{
plot_densities(fits)
}
\arguments{
\item{fits}{the fits from fit_mixtures}
}
\description{
Plot densities of competing mixture fits for diagnostics.
}
\seealso{
fit_mixtures
}
