% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_velo.R
\name{plot_velo}
\alias{plot_velo}
\title{Generate interactive 3D UMAP plot with cones for velocity. Work in progress.}
\usage{
plot_velo(
  txis,
  embed = "UMAP",
  replace = FALSE,
  colr = "velocity_pseudotime",
  flip = FALSE,
  ...
)
}
\arguments{
\item{txis}{a SingleCellExperiment with metadata(txis)$scVelo}

\item{embed}{which reducedDims() element to use for plotting? (UMAP)}

\item{replace}{replace any existing metadata element \code{embedded}? (FALSE)}

\item{colr}{column(s) to use for colors (default: velocity_pseudotime)}

\item{flip}{flip the velocity arrows (ghetto CellRank)? (FALSE)}

\item{...}{optional arguments to pass to plotly::plot_ly()}
}
\value{
\preformatted{       a plotly plot
}
}
\description{
The opacity of the cones is manipulated so that dot colors are visible.
The provided SingleCellExperiment \code{txis} MUST have an element \code{scVelo} in
its metadata, and that element (a SingleCellExperiment) needs to have a
colData column "velocity_pseudotime" unless an argument 'colr' is provided.
(Hence the references internally to "COLORING" rather than "COLOR")
}
