% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_velo_txis.R
\name{process_velo_txis}
\alias{process_velo_txis}
\title{Load some Alevin runs quantified with velocity information}
\usage{
process_velo_txis(
  runs,
  txstub,
  anno = NULL,
  qm = "alevin/quants_mat.gz",
  QC = TRUE,
  HARMONY = FALSE,
  CLUSTER = TRUE,
  DEDUPE = FALSE,
  SCVELO = FALSE,
  ...,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{runs}{a named vector of paths (if unnamed, you'll have problems)}

\item{txstub}{the shared portion of the transcriptome annotation files}

\item{anno}{optional GRanges linking e.g. ENSG transcripts to HGNC symbols}

\item{qm}{the name and path of the Alevin quants matrix file in 'runs'}

\item{QC}{do rudimentary quality control on each Alevin dataset? (TRUE)}

\item{HARMONY}{run Harmony on merged dataset? (FALSE, as harmony is non-BioC)}

\item{CLUSTER}{apply simple Louvain clustering to the result? (TRUE)}

\item{DEDUPE}{de-dupe genes and cells (doublets)? (FALSE unless SCVELO=TRUE)}

\item{SCVELO}{run scVelo on the (perhaps harmonized) result? (FALSE, slow)}

\item{...}{other arguments to pass to velociraptor::scvelo, if any}

\item{BPPARAM}{if running in parallel, provide a MulticoreParam or similar}
}
\value{
\preformatted{   a SingleCellExperiment with 
}
}
\description{
The default assay (which is then split into spliced and unspliced matrices)
is rescaled TPM, which is to say, abundance in transcripts per million,
scaled by the number of fragments/transcripts observed in each cell.
}
\details{
We have found when comparing single-cell protocols against bulk and against
each other that this allows for both correction of sampling bias, and also
the use of count-like representations instead of purely proportional TPMs.
If you don't have a good reason to pass something other than 'abundance' in
the optional arguments to 'asy', we would suggest leaving it as is.

The results of this function can be further QC'ed and/or batch corrected,
for example with Harmony or sva::ComBat, then passed to velociraptor::scvelo
and cellrank::cellrank for data-driven trajectory/cell-fate inference. We
are investigating whether a factorization approach may allow for sensible
batch effect correction while respecting the relationships between spliced
and unspliced transcript abundance, but have not yet finalized it.

See also plotVelocity for a relatively straightforward exploratory plot
based on UMAP coordinates and a downsampled subset of cells from the
SingleCellExperiment this function and velociraptor::scvelo yields.
}
